py_quill/requirements.txt
adk
anthropic
bs4
firebase-functions
functions-framework
google-adk
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
google-cloud-aiplatform
google-cloud-firestore
google-cloud-logging
google-cloud-secret-manager
google-cloud-storage
google-cloud-texttospeech
google-cloud-vision
google-genai
more_itertools
openai
pandas
Pillow
pytest
pytest-cov
pytest>=7.3.1
requests
tabulate
tqdm
urllib3
vertexai
